:truck: Move sanitization functions to shared
