+++ src/interactors/LearningObjectInteractor.ts
@@ -2,7 +2,7 @@
 import * as stopword from 'stopword';
 import { reportError } from '../drivers/SentryConnector';
 import { processMultipartUpload } from '../FileManager/FileInteractor';
-import { sanitizeObject, sanitizeText } from '../functions';
+import { sanitizeObject, sanitizeText } from '../shared/functions';
 import {
   LearningObjectQuery,
   QueryCondition,
